+++ tenpy/tools/misc.py
@@ -343,3 +343,22 @@ def any_nonzero(params, keys, verbose_msg=None):
                     print(str(k) + " has nonzero entries")
                 return True
     return False
+
+def add_with_None_0(a, b):
+    """Return ``a + b``, treating `None` as zero.
+
+    Parameters
+    ----------
+    a, b :
+        The two things to be added, or ``None``.
+
+    Returns
+    -------
+    sum :
+        ``a + b``, except if `a` or `b` is `None`, in which case the other variable is returned.
+    """
+    if a is None:
+        return b
+    if b is None:
+        return a
+    return a + b