+++ CHANGELOG.rst
@@ -39,6 +39,7 @@ Added
 - request #8: Allow shift in boundary conditions of :class:`~tenpy.models.model.CouplingModel`.
 - Allow to use state labels in :meth:`tenpy.networks.mps.from_product_state`.
 - :class:`tenpy.models.model.CouplingMPOModel` structuring the default initialization of most models.
+- :meth:`tenpy.models.model.NearestNeighborModel.calc_H_MPO_from_bond` for conversion of H_bond into H_MPO.
 
 Changed
 ^^^^^^^